+++ Script_Sprite.h
@@ -13,10 +13,10 @@ class Script_Sprite : public Script_Billboard {
 private:
 	bool  mIsEndAnimation{ false };		// 애니메이션이 종료되었는가? (마지막 프레임에 도달했는가?)
 
-	float mRows{};			// 행 개수
-	float mRow{};			// 현재 행
-	float mCols{};			// 열 개수
-	float mCol{};			// 현재 열
+	int mRows{};	// 행 개수
+	int mRow{};		// 현재 행
+	int mCols{};	// 열 개수
+	int mCol{};		// 현재 열
 	
 	float mSpeed{ 0.01f };	// 애니메이션 속도
 	float mElapsedTime{};	// 애니메이션 실행 경과시간
